requirements.txt
streamlit
sentence-transformers
faiss-cpu
transformers
torch
nltk
tqdm
numpy
accelerate
watchdog
matplotlib